sdk: Make deplint pass
Basically involves moving some top level types to the SDK and exporting
them from there.

Change-type: patch
