+++ lib/sdk/auth.ts
@@ -15,8 +15,8 @@
  */
 
 import * as Bluebird from 'bluebird';
-import { Card } from '../types';
 import { SDKInterface } from './index';
+import { Card } from './types';
 import { debug } from './utils';
 
 // A regex used to test that a string contains only alphanumeric characters and
